+++ sdk/resourcemanager/recoveryservices/armrecoveryservices/registeredidentities_client.go
@@ -57,7 +57,7 @@ func NewRegisteredIdentitiesClient(subscriptionID string, credential azcore.Toke
 
 // Delete - Unregisters the given container from your Recovery Services vault.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-04-01
+// Generated from API version 2022-08-01
 // resourceGroupName - The name of the resource group where the recovery services vault is present.
 // vaultName - The name of the recovery services vault.
 // identityName - Name of the protection container to unregister.
@@ -102,7 +102,7 @@ func (client *RegisteredIdentitiesClient) deleteCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-04-01")
+	reqQP.Set("api-version", "2022-08-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }