+++ sdk/resourcemanager/recoveryservices/armrecoveryservices/replicationusages_client.go
@@ -56,8 +56,7 @@ func NewReplicationUsagesClient(subscriptionID string, credential azcore.TokenCr
 }
 
 // NewListPager - Fetches the replication usages of the vault.
-// If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-04-01
+// Generated from API version 2022-08-01
 // resourceGroupName - The name of the resource group where the recovery services vault is present.
 // vaultName - The name of the recovery services vault.
 // options - ReplicationUsagesClientListOptions contains the optional parameters for the ReplicationUsagesClient.List method.
@@ -103,7 +102,7 @@ func (client *ReplicationUsagesClient) listCreateRequest(ctx context.Context, re
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-04-01")
+	reqQP.Set("api-version", "2022-08-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil